+++ source/CMakeLists.txt
@@ -1,7 +1,3 @@
-add_executable(main main.cpp)
-
-set_target_properties(
-    main
-    PROPERTIES
-    RUNTIME_OUTPUT_DIRECTORY ${TARGET_PATH}
-)
+add_subdirectory(distance)
+add_subdirectory(bruteforce)
+add_subdirectory(nnhnsw)
