+++ source/bruteforce/CMakeLists.txt
@@ -1,0 +1,12 @@
+add_library(bruteforce "")
+target_sources(bruteforce
+  PRIVATE
+    ${CMAKE_CURRENT_LIST_DIR}/bruteforce.cpp
+  PUBLIC
+    ${CMAKE_CURRENT_LIST_DIR}/bruteforce.hpp
+  )
+target_include_directories(bruteforce
+  PUBLIC
+      ${CMAKE_CURRENT_LIST_DIR}
+  )
+target_link_libraries(bruteforce PRIVATE distance)
